+++ webServer/project/Webserver/BigTeta/urls.py
@@ -2,6 +2,7 @@ from django.urls import path
 from django.views.generic.base import TemplateView
 
 from . import views
+from . import api
 app_name = 'BigTeta'
 urlpatterns = [
     path('', views.showFiles, name='showVideos'),
@@ -10,7 +11,7 @@ urlpatterns = [
     path('register', views.register, name='register'),
     path('logout', views.logout_user, name='logout'),
     path('show', views.show_video, name='showVideo'),
-    path('getvideos', views.get_video_by_title, name='getVideos'),
+    path('getvideos', api.get_video_by_title, name='getVideos'),
     path('showFiles', views.showFiles, name='showFiles'),
     path('upload', views.upload, name='upload'),
     path('prueba', views.prueba, name='prueba'),
